app (deoricated)/requirements.txt
aiopg==1.4.0
bcrypt==4.1.2
celery==5.3.6
loguru==0.7.2
passlib==1.7.4
psycopg2-binary==2.9.9
pydantic==2.6.3
pyhumps==3.8.0
Pyrogram==1.3.0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
TgCrypto==1.2.5
uvicorn==0.27.1
SQLAlchemy==1.4.41
fastapi==0.110.0
pyrogram-aiopg-storage==0.3


